Make the users.json file permanent
Instead of having the setup potentially copy over the example user file,
just ask the user to set it up himself. This stops mistakes from
happening with `--force`.

Also mount the file into the container instead of copying it, so that an
update just needs a container restart and not a rebuild. Hence, remove
the rebuilding from the start script and add it as a developer-intended
command to the npm scripts (since the only reason to rebuild would be if
the pushStats scripts is edited).